feat(mise): support Windows
The mise step was gated to Unix. Move it into the shared module and run it
everywhere, since mise works on Windows too.
