+++ src/steps/os/unix.rs
@@ -1,6 +1,8 @@
 use color_eyre::eyre::Context;
+#[cfg(any(target_os = "linux", target_os = "macos"))]
+use color_eyre::eyre::OptionExt;
 use color_eyre::eyre::Result;
-use color_eyre::eyre::{OptionExt, bail, eyre};
+use color_eyre::eyre::{bail, eyre};
 use etcetera::BaseStrategy;
 use ini::Ini;
 use regex::Regex;
@@ -8,13 +10,12 @@ use rust_i18n::t;
 use semver::Version;
 use std::env::home_dir;
 use std::ffi::OsStr;
-use std::io::Write;
+use std::fs;
 use std::os::unix::fs::MetadataExt;
 use std::path::Component;
 use std::path::PathBuf;
 use std::sync::LazyLock;
 use std::{env::var, path::Path};
-use std::{fs, io};
 use tracing::{debug, warn};
 
 use crate::XDG_DIRS;
@@ -867,68 +868,6 @@ pub fn run_asdf(ctx: &ExecutionContext) -> Result<()> {
     }
 
     ctx.execute(&asdf).args(["plugin", "update", "--all"]).status_checked()
-}
-
-pub fn run_mise(ctx: &ExecutionContext) -> Result<()> {
-    let mise = require("mise")?;
-
-    print_separator("mise");
-
-    ctx.execute(&mise).args(["plugins", "update"]).status_checked()?;
-
-    let output = ctx
-        .execute(&mise)
-        .args(["self-update"])
-        .output_checked_with(|_| Ok(()))?;
-    let status_code = output
-        .status
-        .code()
-        .ok_or_eyre("Couldn't get status code (terminated by signal)")?;
-    let stderr = std::str::from_utf8(&output.stderr).wrap_err("Expected output to be valid UTF-8")?;
-    if stderr.contains("cannot update") && status_code == 1 {
-        debug!("Mise self-update not available")
-    } else {
-        // Write the output
-        io::stdout().write_all(&output.stdout)?;
-        io::stderr().write_all(&output.stderr)?;
-        if status_code != 0 {
-            return Err(StepFailed.into());
-        }
-    }
-
-    let mut cmd = ctx.execute(&mise);
-
-    cmd.arg("upgrade");
-
-    if ctx.config().mise_interactive() {
-        cmd.arg("--interactive");
-    }
-
-    if ctx.config().mise_bump() {
-        cmd.arg("--bump");
-    }
-
-    if ctx.config().mise_silent() {
-        cmd.arg("--silent");
-    }
-
-    if ctx.config().mise_quiet() {
-        cmd.arg("--quiet");
-    }
-
-    if ctx.config().mise_verbose() {
-        cmd.arg("--verbose");
-    }
-
-    if ctx.config().yes(Step::Mise) {
-        cmd.arg("--yes");
-    }
-
-    if ctx.config().mise_jobs() != 4 {
-        cmd.args(["--jobs", &ctx.config().mise_jobs().to_string()]);
-    }
-
-    cmd.status_checked()
 }
 
 pub fn run_home_manager(ctx: &ExecutionContext) -> Result<()> {
